backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
firebase-admin==6.5.0
google-generativeai==0.8.3
google-cloud-tasks==2.16.3
google-auth==2.36.0
google-api-python-client==2.154.0
pydantic==2.10.0
python-multipart==0.0.12
httpx==0.27.2
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
black==24.10.0
mypy==1.13.0
